requirements.txt
asgiref==3.5.2
certifi==2021.10.8
Django==4.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8